+++ api/comments/iterator.go
@@ -20,7 +20,7 @@ type CommentIterator interface {
 	Next(ctx context.Context) ([]*types.Comment, error)
 }
 
-func commentListIterator(ctx context.Context, s *Service, postID string, limit int) CommentIterator {
+func commentListIterator(s *Service, postID string, limit int) CommentIterator {
 	fetch := func(ctx context.Context, token iterator.PageToken) ([]*types.Comment, iterator.PageToken, bool, error) {
 		result, err := s.getCommentList(ctx, postID, token.Cursor, limit)
 		if err != nil {