+++ api/notifications/dto.go
@@ -1,0 +1,10 @@
+package notifications
+
+import (
+	"github.com/slipynil/itd-go/types"
+)
+
+type notificationResponse struct {
+	Notifications []*types.Notification `json:"notifications"`
+	HasMore       bool                  `json:"hasMore"`
+}